--- src/app/dashboard/page.tsx
@@ -1,24 +1,0 @@
-import AppSidebar from '@/components/AppSidebar';
-import Navbar from '@/components/Navbar';
-import { SidebarProvider } from '@/components/ui/sidebar';
-import React from 'react';
-
-const DashboardPage = ({
-  children,
-  defaultOpen,
-}: {
-  children: React.ReactNode;
-  defaultOpen: boolean;
-}) => {
-  return (
-    <SidebarProvider defaultOpen={defaultOpen}>
-      <AppSidebar />
-      <main className="w-full">
-        <Navbar />
-        <div className="px-4">{children}</div>
-      </main>
-    </SidebarProvider>
-  );
-};
-
-export default DashboardPage;
