+++ src/lib/actions/loginSetup.ts
@@ -1,8 +1,12 @@
 'use server';
 
-import { signIn } from './authSetup';
+import { signIn, signOut } from './authSetup';
 
 export const doSocialLogin = async (formData: FormData) => {
   const action = formData.get('action');
   await signIn(action as string, { redirectTo: '/leave' });
+};
+
+export const doLogout = async () => {
+  await signOut({ redirectTo: '/' });
 };
